Practice char.c created

+++ typecasting.c
@@ -1,5 +1,12 @@
 #include <stdio.h>
 
+
+/**********************
+ * Author: Udayraj
+ * Description: App to practice type casting and commenting in C
+ * Date: 10/08/2020
+ * **********************/
+
 int main()
 {
 	int eggs;
@@ -7,7 +14,7 @@ int main()
 	printf("The number of eggs for the day: ");
 	scanf("%i" , &eggs);
 
-	double dozen = (double) eggs/12; //Type casting double allows the answer of the datatype of the expression to change
+	float dozen = (double) eggs/12; //Type casting double allows the answer of the datatype of the expression to change
 
 	printf(" You have %f dozen eggs.\n", dozen);
